single configureStore
