--- app/shared/store/configureStore.production.js
@@ -1,16 +1,0 @@
-// see dev version!!!
-
-
-// import { createStore, applyMiddleware } from 'redux';
-// import thunk from 'redux-thunk';
-// import { hashHistory } from 'react-router';
-// import { routerMiddleware } from 'react-router-redux';
-// import rootReducer from '../reducers';
-//
-// const router = routerMiddleware(hashHistory);
-//
-// const enhancer = applyMiddleware(thunk, router);
-//
-// export default function configureStore(initialState) {
-//   return createStore(rootReducer, initialState, enhancer);
-// }
